requirements.txt
arrow==1.4.0
asgiref==3.10.0
binaryornot==0.4.4
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2.8
django-appconf==1.2.0
django-browser-reload==1.21.0
django-debug-toolbar==6.1.0
django-redis==6.0.0
django-select2==8.4.3
django-tailwind==4.4.1
django-unfold==0.69.0
gunicorn==21.2.0
idna==3.11
isort==7.0.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.2.0
six==1.17.0
slugify==0.0.1
sqlparse==0.5.3
stripe==13.2.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
